+++ domain/src/main/java/ru/dpopkov/knownics/domain/answer/SourceDetails.java
@@ -27,4 +27,22 @@ public class SourceDetails extends ModifiableEntity {
     public void setDetails(String details) {
         this.details = details;
     }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof SourceDetails)) return false;
+
+        SourceDetails details1 = (SourceDetails) o;
+
+        if (!getSource().equals(details1.getSource())) return false;
+        return getDetails().equals(details1.getDetails());
+    }
+
+    @Override
+    public int hashCode() {
+        int result = getSource().hashCode();
+        result = 31 * result + getDetails().hashCode();
+        return result;
+    }
 }