+++ domain/src/main/java/ru/dpopkov/knownics/domain/question/Category.java
@@ -30,4 +30,19 @@ public class Category extends ModifiableEntity {
     public void setDescription(String description) {
         this.description = description;
     }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (!(o instanceof Category)) return false;
+
+        Category category = (Category) o;
+
+        return getName() != null ? getName().equals(category.getName()) : category.getName() == null;
+    }
+
+    @Override
+    public int hashCode() {
+        return getName() != null ? getName().hashCode() : 0;
+    }
 }