+++ history.md
@@ -3,6 +3,7 @@ History of implementation
 
 ##### 2021-12-26
 * Add module services. Add CategoryService, CategoryServiceImpl.
+* Implement equals and hashCode for entities. Closes #9. 
 ##### 2021-12-25
 * Add module repository-map. Implement Map based repositories. Closes #5.
 ##### 2021-12-10